backend/requirements.txt
fastapi
uvicorn
python-multipart
psycopg2-binary
gunicorn
supabase
google-generativeai
google-cloud-aiplatform